auto insert pro code
