+++ src/layouts/BasicLayout.tsx
@@ -1,6 +1,5 @@
 import { ConnectState, ConnectProps } from '@/models/connect';
 import RightContent from '@/components/GlobalHeader/RightContent';
-import CopyBlock from '@/components/CopyBlock';
 import { connect } from 'dva';
 import React, { useState } from 'react';
 import logo from '../assets/logo.svg';
@@ -11,30 +10,26 @@ import {
   BasicLayoutProps as BasicLayoutComponentsProps,
   MenuDataItem,
   Settings,
-  SettingDrawer,
 } from '@ant-design/pro-layout';
 import Link from 'umi/link';
-import { isAntDesignProOrDev } from '@/utils/utils';
-
 export interface BasicLayoutProps extends BasicLayoutComponentsProps, ConnectProps {
-  breadcrumbNameMap: { [path: string]: MenuDataItem };
+  breadcrumbNameMap: {
+    [path: string]: MenuDataItem;
+  };
   settings: Settings;
 }
-
 export type BasicLayoutContext = { [K in 'location']: BasicLayoutProps[K] } & {
-  breadcrumbNameMap: { [path: string]: MenuDataItem };
+  breadcrumbNameMap: {
+    [path: string]: MenuDataItem;
+  };
 };
-
 /**
  * use Authorized check all menu item
  */
+
 const menuDataRender = (menuList: MenuDataItem[]): MenuDataItem[] => {
   return menuList.map(item => {
-    const localItem = {
-      ...item,
-      children: item.children ? menuDataRender(item.children) : [],
-    };
-
+    const localItem = { ...item, children: item.children ? menuDataRender(item.children) : [] };
     return Authorized.check(item.authority, localItem, null) as MenuDataItem;
   });
 };
@@ -44,57 +39,52 @@ const BasicLayout: React.FC<BasicLayoutProps> = props => {
   /**
    * constructor
    */
+
   useState(() => {
-    dispatch!({ type: 'user/fetchCurrent' });
-    dispatch!({ type: 'settings/getSetting' });
+    dispatch!({
+      type: 'user/fetchCurrent',
+    });
+    dispatch!({
+      type: 'settings/getSetting',
+    });
   });
   /**
    * init variables
    */
+
   const handleMenuCollapse = (payload: boolean) =>
-    dispatch!({ type: 'global/changeLayoutCollapsed', payload });
+    dispatch!({
+      type: 'global/changeLayoutCollapsed',
+      payload,
+    });
 
   return (
-    <>
-      <BasicLayoutComponents
-        logo={logo}
-        onCollapse={handleMenuCollapse}
-        menuItemRender={(menuItemProps, defaultDom) => {
-          return <Link to={menuItemProps.path}>{defaultDom}</Link>;
-        }}
-        breadcrumbRender={(routers = []) => {
-          return [
-            {
-              path: '/',
-              breadcrumbName: formatMessage({
-                id: 'menu.home',
-                defaultMessage: 'Home',
-              }),
-            },
-            ...routers,
-          ];
-        }}
-        menuDataRender={menuDataRender}
-        formatMessage={formatMessage}
-        rightContentRender={rightProps => <RightContent {...rightProps} />}
-        {...props}
-        {...settings}
-      >
-        {children}
-      </BasicLayoutComponents>
-      {isAntDesignProOrDev() && (
-        <SettingDrawer
-          settings={settings}
-          onSettingChange={config =>
-            dispatch!({
-              type: 'settings/changeSetting',
-              payload: config,
-            })
-          }
-        />
-      )}
-      {isAntDesignProOrDev() && <CopyBlock url={location!.pathname} />}
-    </>
+    <BasicLayoutComponents
+      logo={logo}
+      onCollapse={handleMenuCollapse}
+      menuItemRender={(menuItemProps, defaultDom) => {
+        return <Link to={menuItemProps.path}>{defaultDom}</Link>;
+      }}
+      breadcrumbRender={(routers = []) => {
+        return [
+          {
+            path: '/',
+            breadcrumbName: formatMessage({
+              id: 'menu.home',
+              defaultMessage: 'Home',
+            }),
+          },
+          ...routers,
+        ];
+      }}
+      menuDataRender={menuDataRender}
+      formatMessage={formatMessage}
+      rightContentRender={rightProps => <RightContent {...rightProps} />}
+      {...props}
+      {...settings}
+    >
+      {children}
+    </BasicLayoutComponents>
   );
 };
